+++ src/main/java/com/gsm/blabla/auth/application/AppleService.java
@@ -5,9 +5,9 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.gsm.blabla.global.exception.GeneralException;
 import com.gsm.blabla.global.response.Code;
 import com.gsm.blabla.auth.dao.AppleAccountRepository;
-import com.gsm.blabla.jwt.dto.AppleAccountDto;
-import com.gsm.blabla.jwt.dto.ApplePublicKeyDto;
-import com.gsm.blabla.jwt.dto.AppleTokenDto;
+import com.gsm.blabla.auth.dto.AppleAccountDto;
+import com.gsm.blabla.auth.dto.ApplePublicKeyDto;
+import com.gsm.blabla.auth.dto.AppleTokenDto;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.SignatureAlgorithm;
