+++ src/main/java/com/gsm/blabla/auth/application/AuthService.java
@@ -10,9 +10,9 @@ import com.gsm.blabla.jwt.dao.JwtRepository;
 import com.gsm.blabla.auth.domain.AppleAccount;
 import com.gsm.blabla.auth.domain.GoogleAccount;
 import com.gsm.blabla.jwt.domain.Jwt;
-import com.gsm.blabla.jwt.dto.AppleAccountDto;
-import com.gsm.blabla.jwt.dto.AppleTokenDto;
-import com.gsm.blabla.jwt.dto.GoogleAccountDto;
+import com.gsm.blabla.auth.dto.AppleAccountDto;
+import com.gsm.blabla.auth.dto.AppleTokenDto;
+import com.gsm.blabla.auth.dto.GoogleAccountDto;
 import com.gsm.blabla.jwt.dto.JwtDto;
 import com.gsm.blabla.jwt.dto.TokenRequestDto;
 import com.gsm.blabla.member.dao.MemberRepository;
