+++ src/main/java/com/gsm/blabla/auth/dto/ApplePublicKeyDto.java
@@ -1,4 +1,4 @@
-package com.gsm.blabla.jwt.dto;
+package com.gsm.blabla.auth.dto;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import java.util.ArrayList;
