+++ src/main/java/com/gsm/blabla/auth/dto/AppleTokenDto.java
@@ -1,4 +1,4 @@
-package com.gsm.blabla.jwt.dto;
+package com.gsm.blabla.auth.dto;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
